142682 - changed IsEnabled property datatype AB#142682
