+++ UKHO.FmEssFssMock.API/UKHO.FmEssFssMock.API/Models/Bess/Tag.cs
@@ -1,0 +1,13 @@
+﻿using Newtonsoft.Json;
+
+namespace UKHO.FmEssFssMock.API.Models.Bess
+{
+    public class Tag
+    {
+        [JsonProperty("key")]
+        public string Key { get; set; }
+
+        [JsonProperty("value")]
+        public string Value { get; set; }
+    }
+}
